+++ packages/flutter/test/widgets/media_query_test.dart
@@ -4,6 +4,7 @@
 
 import 'dart:ui' show Brightness;
 
+import 'package:flutter/gestures.dart';
 import 'package:flutter/widgets.dart';
 import 'package:flutter_test/flutter_test.dart';
